# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "SpiderX is a private tracker server.",
        "title": "SpiderX API",
        "contact": {
            "name": "EndPot",
            "url": "https://github.com/endpot"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "0.0.1"
    },
    "host": "localhost:8888",
    "basePath": "/",
    "paths": {
...
        "/chats": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取群聊消息列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "获取群聊消息列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 15,
                        "description": "每页数量",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.PageResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "allOf": [
                                                    {
                                                        "$ref": "#/definitions/chat.Chat"
                                                    },
                                                    {
                                                        "type": "object",
                                                        "properties": {
                                                            "user": {
                                                                "$ref": "#/definitions/chat.User"
                                                            }
                                                        }
                                                    }
                                                ]
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数异常",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "没有操作权限",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "没有对象",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建群聊",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "创建群聊",
                "parameters": [
                    {
                        "description": "创建聊天消息请求参数",
                        "name": "chat",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/chat.CreateChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/chat.Chat"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "user": {
                                                            "$ref": "#/definitions/chat.User"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数异常",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "没有操作权限",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "没有对象",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    }
                }
            }
        },
        "/chats/{chat_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取群聊消息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "获取群聊消息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "群聊ID",
                        "name": "chat_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/chat.Chat"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "user": {
                                                            "$ref": "#/definitions/chat.User"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数异常",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "没有操作权限",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "没有对象",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "删除群聊消息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "删除群聊消息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "群聊ID",
                        "name": "chat_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "请求成功"
                    },
                    "400": {
                        "description": "请求参数异常",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "没有操作权限",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "没有对象",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "更新群聊消息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "更新群聊消息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "群聊ID",
                        "name": "chat_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新群聊请求参数",
                        "name": "chat",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/chat.UpdateChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/chat.Chat"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "user": {
                                                            "$ref": "#/definitions/chat.User"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数异常",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "没有操作权限",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "没有对象",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "chat.Chat": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "聊天内容",
                    "type": "string",
                    "example": "This is a notice"
                },
                "created_at": {
                    "description": "创建时间戳（秒）",
                    "type": "integer",
                    "example": 1591974665
                },
                "id": {
                    "description": "ID",
                    "type": "integer",
                    "example": 1
                },
                "updated_at": {
                    "description": "更新时间戳（秒）",
                    "type": "integer",
                    "example": 1591974665
                },
                "user": {
                    "description": "用户",
                    "type": "object"
                }
            }
        },
        "chat.CreateChatRequest": {
            "type": "object",
            "required": [
                "content"
            ],
            "properties": {
                "content": {
                    "description": "聊天内容",
                    "type": "string",
                    "example": "This is a chat"
                }
            }
        },
        "chat.UpdateChatRequest": {
            "type": "object",
            "required": [
                "content"
            ],
            "properties": {
                "content": {
                    "description": "聊天内容",
                    "type": "string",
                    "example": "This is a chat"
                }
            }
        },
        "chat.User": {
            "type": "object",
            "properties": {
                "display_name": {
                    "description": "用户名",
                    "type": "string",
                    "example": "spider"
                },
                "id": {
                    "description": "用户ID",
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "response.ErrResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "integer",
                    "example": 500
                },
                "message": {
                    "description": "消息",
                    "type": "string"
                }
            }
        },
        "response.PageMeta": {
            "type": "object",
            "properties": {
                "page": {
                    "description": "页码",
                    "type": "integer",
                    "example": 1
                },
                "per_page": {
                    "description": "每页数量",
                    "type": "integer",
                    "example": 15
                },
                "total": {
                    "description": "总数",
                    "type": "integer",
                    "example": 100
                }
            }
        },
        "response.PageResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "description": "数据",
                    "type": "object"
                },
                "meta": {
                    "description": "附加信息",
                    "type": "object",
                    "$ref": "#/definitions/response.PageMeta"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "description": "数据",
                    "type": "object"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}